class-39/demo/cookie-stand-api/cookie_stands/serializers.py
from rest_framework import serializers
from .models import CookieStand


class CookieStandSerializer(serializers.ModelSerializer):
    class Meta:
        model = CookieStand
        fields = "__all__"
