class-39/demo/cookie-stand-api/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.4
Django==4.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
fire==0.4.0
gunicorn==20.1.0
idna==3.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
tomli==2.0.1
urllib3==1.26.8
whitenoise==5.3.0
